Déplacer les types StatistiquesX dans API Pitchou
